[host-target-testing] Support lazily fetching blobs
Some of the OTA test steps do not need all the blobs, like when we pave N-1
onto a device, then validate with sl4f. This introduces a BlobFetchMode,
where if we get a repository with LazilyFetchBlobs, we will only download
blobs when they're requested by the test device. This can half the artifacts
we need to download when getting the old version.

To go even further, we could be even smarter and compare the blobs from
N-1 and N, and only prefetch the blobs that are only in N, but we can address
that optimization down the road.

Change-Id: I4d28694a72fa0b3341cc9fc71e4ec84f1668dd0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512311
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>