[blobfs] Clean up paged BlobLoader path.
Separates out the result of loading a paged vs unpaged blob. The unpaged
path is unchanged.

The paged path is changed to return the PagerInfo and BlobLayout so the
caller (Blob in this case) can subsequently create the vmo and mapping.
This eliminates the need for the vmo creation callback which
significantly simplifies the structure of thise code.

This eliminates the vmo handle duplication in the new (unused) pager
code path which is the source of bugs.

Moves handling of null blobs out of the BlobLoader and into its caller
(the Blob).

There should be no observable behavior change.

Change-Id: I8d7fb4847307fec9ea79862bb160f020554c5953
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512862
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>