[timekeeper] Refactor EstimateUpdated diagnostics event.
In the future we'll have both events from the kalman filter estimate
and from the frequency estimate algorithm so "EstimateUpdate" was not
sufficiently precise. Renamed to "KalmanFilterUpdae".

This also moves the content from a single "offset" field (that
assumed the utc line was parallel to the monotonic line) to a pair of
monotonic/utc which fully specifies a point on the line. Changes to
the frequency will be recorded separately in a different event type
and so don't need to be included.

Bug: 56868
Change-Id: Ibccf1d6f27a53bb3f90d26cfa29d24fb69c2ee34
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511041
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>