[build] Assert is_host in golden_test.gni
This CL makes the golden_test template `assert(is_host)`. This would
have saved confusion recently, when a new golden test setup was not
working because the `($host_toolchain)` was missing.

Doing this required reorganizing call sites so that the golden_test
invocation is inside an `if (is_host)` block. Following the
recommendation of build folks, I've structured host-tool BUILD.gn files
to place most things in one big if-statement, with groups at the bottom
that redirect to $host_toolchain.

Change-Id: Id1e747d322ed5bcfa64060463ca8d91db9e1938a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513324
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>