[Bind] Add device properties fallback
The stored device properties might not contain the
protocol ID or autobind data. As such, when matching
a driver to a device, the protocol ID and autobind
values need to be added to the device properties.

Bug: 74316
Testing: Added unit tests in binding_v2_tests.cc
Change-Id: I2d55bc79ac8db507216b2158988db0e8809f6909
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514108
Commit-Queue: Sarah Chan <spqchan@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>