[blobfs] Improve test setup.
The new pager requires a Vfs which is a more involved setup procedure
than most of the tests were doing. This adds a new test helper class to
manage Blobfs setup and running, and uses it in a numbe of tests.

Remove CreateTest.ValidSuperblock because it was just running the
standard test setup code.

There is no non-test change.

Change-Id: I4d23bcbe6fbfa721c2895eadf4f106a0d406edd2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515321
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>