[fidlc] Allow COMPILE_AND_CONVERT tests for libraries with dependencies
Several new conversion validating macros have been added:
* ASSERT_COMPILED_AND_CONVERT_INTO takes an uninitialized TestLibrary
and populates it with the result of compiling a converted file. This is
useful for converting a library that will be consumed as a dependency of
another library.
* ASSERT_COMPILED_AND_CLONE_INTO is identical to the above, except that
it does not convert the second library (ie, it clones it instead). This
is necessary because we will need two copies of the dependent library:
one to successfully complete the pre-conversion compilation of the
target library, and one to use as an unconverted dependency for its
converted version.
* ASSERT_COMPILED_AND_CONVERT_WITH_DEP allows a library to be converted
with a dependency, generated from one of the two macros above.
* ASSERT_COMPILED_AND_CONVERT_WITH_DEP_INTO is not used directly - it is
only called from inside one of ASSERT_COMPILED_AND_CONVERT_INTO,
ASSERT_COMPILED_AND_CONVERT_WITH_DEP, or ASSERT_COMPILED_AND_CONVERT.

Test: fidldev test fidlc
Bug: 72924
Change-Id: I56ea280f70f863e39a05e9ec4266ab3799442a02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513299
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>