[kernel][x86] Add microarchitectural entries for AMD Zen
Add microarchitectural switch for Zen 3. Should be basically identical
to family 17h though we may want to experiment and see if the HLT for
idle is still power savings. Either way, better than treating it as a
default config for the purposes of detecting clocks.

Also add a proper microarchitectural AMD name for the lib/arch enums.

Change-Id: I563e580d9b154ab7c8a0da294a0e74e228922c41
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515733
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>