[kernel][lib/arch] Add Write() convenience method to SysReg.
Add a Write() convenience method to the SysRegDerived class, allowing
many instances of:

  arch::FooReg::Modify([](auto& reg) {
    reg.set_a(true).set_b(true);
  });

to be written instead as:

  arch::FooReg::Read().set_a(true).set_b(true).Write();

which is a slight improvement in readability.

Like the existing Modify convenience method, this only works if mocking
support is not required in the code.

Change-Id: Id3d4a080cbb0257d5465408caf27da91682fc42c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513622
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>