[fidl][llcpp] Do not assume heap allocated ResponseContext
The ResponseContext classes is central to caller-allocated flavors of
async calls - although not yet used in production, it is possible to
allocate the bookkeeping state for an async method through ways
beside the heap (stack alloc, using allocator, etc.), thus performing
an entire async call without any heap allocation.

There is one place in llcpp_message.cc that assumes ResponseContexts are
always allocated on the heap, and tries to delete it on error. This CL
fixes that to properly release ownership by calling |OnError|.

Improved the documentation on the lifecycle and ownership of
ResponseContext.

TEST: added one in integration
Change-Id: I8eea545c96309404ab54331c10e533c9233c11fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514113
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>