[bsdsocket_test] Wait for the enqueue of passive connection
This test expects to accept a connection that is reset after it was
enqueued. Deflake the test by ensuring that the connection is reset
only after a successful enqueue. When the listener is using syncookies,
the RST could arrive before the final ACK of the handshake is processed,
causing it to get dropped by the listener.

Fixed: 74478
Change-Id: I1e8e087d78ecaf80b8c349be806523da8a633634
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516147
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Mithun Iyer <iyerm@google.com>