[fint] Emit ninja targets from `fint build`
I realized I forgot to start setting this field when I added it to the
fint build artifacts proto. Fortunately it's not load-bearing yet, but
will be after fxrev.dev/515584 lands, so this needs to land first.

This is another sign that this function really needs unit tests, but
deleting all the recipe logic that's now duplicated in fint is a
higher-priority task that this fix unblocks. Once that cleanup is done
I'll circle back and add tests here.

Bug: 67861
Change-Id: I11c40ebc3fd65447dc526b52b969a4ae60b92822
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516223
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>