[audio_core] Dest gain ramping
When a stream is ramped at the same time that its usage changes, there
is a collision in the Gain object's ramp values, leading to inconsistent
behavior. These should be accounted for separately instead of being
combined into the source_gain side of Gain, while the dest_gain side
sits unused.  One reason dest_gain was not used was that it did not
implement ramping.

This CL adds SetDestGainWithRamp to the Gain object, as well as the
source_mute_ boolean. The two gain components (corresponding to the
object's source and dest) continue to be entirely independent, and
are now both orthogonally rampable.

AudioCore does not yet take advantage of this capability; that will
be in a subsequent CL.

New unittests added to audio_mixer_unittests

fx test -o audio_mixer_unittests
Bug: 74275

Change-Id: Ied106cb4ed7d1bf24124d56d1d879666f71eb341
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514659
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>