[kernel][arm64] Remove code saving the percpu pointer around SMC calls
Previously, the percpu pointer (x15) needed to be saved acround SMC
calls because it was possible that it was trashed in the call. This also
meant we needed to disable preemption which turned out to be a subtle
bug.

After moving the kernel's percpu pointer away from x15 to x20, which is
callee and SMC saved, we can remove the need to save and restore this
register.

Also no longer need to save the shadow call pointer (x18) across SMC calls
because it was always callee saved.

Removed the lockup detector logic in the high level smc call wrapper
logic. Filed bug to track future change to add a new version of that
functionality back.

Bug: 69037
Related bug: 74553

Change-Id: Ie877a9ee698287e2a006c3879e3bd6f97de17f27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513645
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: RJ Ascani <rjascani@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>