[kernel][arm64] slight optimization for hand written assembly in context switch
As a general optimization, if pushing or popping a bunch of words
on the stack or other structure, it's better to adjust the base first
and then use a series of load/stores with an immediate offset.

Adjust the core context switch routine to do this.

References:
cortex-a55 optimization guide, for simpler, dual issue cores.
cortex-a72 and above optimize guides, for out of order cores.

Change-Id: I82f0caabfc40ad11720bb51f2214625da73cfaa9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514042
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>