[stress_tests] Swallow all errors in storage stress tests
When an actor in a storage stress test gets an unexpected
status code, assume that it is because of an intentional crash.

The environment will verify that an intentional crash occurred
during reset. If this is not the case, a panic will occur.

This should help with the whack-a-mole situation we're having
where we have to continuously allowlist more status codes
that show up during an intentional crash.

Multiply: blobfs-short: 100
Multiply: minfs-short: 100
Multiply: fvm-short: 100
Test: fx test blobfs-crash
      fx test minfs-crash
      fx test fvm-crash
Change-Id: Ie2762c97d9a126b6dcda1e821afb31fe675ec69e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516280
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>