[blobfs] Maintain a separate non-paged Blob vmo
In the old pager, the vmo_ on a Blob is either a paged VMO for paged
blobs, or holds the data directly for unpaged blobs. In the new pager,
the vmo_ is created by the VFS and it will always be registered as a
paged blob. This means we can't write directly into it.

This separates out a vmo for the case where we need an unpaged buffer.
Either this or the paged vmo_ will be valid when the blob data is
loaded.

This will allow us to support paging from all blobs in the future, even
ones using old compression formats.

This will also allow us to support reading from blobs that were just
written but not yet closed if we so choose in the future.

There should be no behavior change.

Change-Id: I00db33727c21cd547dc9814d92790ec9987bf554
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515602
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>