[rust] Use backtrace-sys mirror
Update and use the backtrace-sys tiny mirror instead of duplicating it
with a vendored copy. This crate is a wrapper around libbacktrace for
use with the backtrace crate. The library isn't intended to be used with
Fuchsia anyway:
https://github.com/rust-lang/backtrace-rs/commit/6bc5be152e8b8b31e8c3236326c60f5d8485c942

This removes the copy of Tom Sawyer, which isn't in line with our policy
on inclusive language.

Change-Id: If0f4727d27e6faa7697a5cf25cdfe5ce238885e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511725
Commit-Queue: Tyler Mandry <tmandry@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>