[fdio] Add fdio_ns_is_bound
Add a function for querying whether a given path is bound to a remote
directory in a local namespace. This function was requested by a
customer to check whether a directory capability was provided in a way
that is guaranteed to be non-blocking.

Bug: 74495
Change-Id: I2b39a71faea449d23a98fd5f0ed62b177b0ba4cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515730
API-Review: Adam Barth <abarth@google.com>
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>