[bringup] Add virtcon2
This is the initial implementation of a Carnelian based
virtual console. See README.md for more details on how
it is different from legacy Virtcon and the transition
plan.

This initial version is not meant to be a full replacement
for legacy Virtcon. It demonstrates efficient use of
vector graphics, and provides a dumb implementation of the
session manager service in order to demonstrate how
functionality in the Terminal app can be extracted into
libraries for code reuse.

Test: fx test virtual_console_tests
Bug: 74628
Change-Id: I82e31fd055f130100d94da03ae4196aae0f3a03c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511880
Reviewed-by: Rob Tsuk <robtsuk@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: David Reveman <reveman@google.com>