[kernel][lib/arch] Make arch::InitializeBootCpuid C++-friendly
This function was only originally meant to be called in assembly. Now
that we have a C++ use-case for it, we update it to ensure that ebx/rbx
- caller-saved - are indeed saved (previously they were being thrashed).
While we're at it, we add CFI metadata. Note: the function is still
buggy on repeated calls after this change; a follow-up will address
that.

Bug: 74527
Change-Id: I53e7b8f574124c328d1747f5e221536af6fc6da1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516061
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>