[ffx] Add `ssh-key` flag to `ffx target flash`
The ssh param is a simplified version of the OEM stage param
specifically for the `add-staged-bootloader-file ssh.authorized_keys`
OEM command. If the key is not supplied, it will now check the config
for the ssh public key to ensure ffx can connect after flashing.

It will also look for a default "fuchsia" product if no product is
supplied via the CLI and there are multiple products found.

MULTIPLY: ffx_flash_lib_test
Fixed: 74314, 74307, 74313

Change-Id: I7cc74f182e27937f49ecf223de0c3928e32996a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515610
Commit-Queue: Matthew Boetger <boetger@google.com>
Reviewed-by: Amit Uttamchandani <amituttam@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>