[blobfs] Stop storing the inode in the blob
Storing the inode in the blob as we were was prone to error because
after writing a blob, it isn't in sync with what has been written; only
some of the fields can be trusted. It's also a bit wasteful on memory.

Fixed: 74799
Change-Id: I5f7787c9262c90d0cb2222b8651c1634e05d2ad5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517440
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Brett Wilson <brettw@google.com>