[benchmarks] Remove step that rounds values to ints when summarising data
The rounding was done on the assumption that we don't care about
sub-nanosecond variation in test results, but this has turned out to
be not quite right for results with low variation.

These result values are themselves averages, so the fractions can be
meaningful, and it's useful to be able to visualise the fractional
values in graphs (such as in Chromeperf) rather than seeing the metric
distributed across a small number of integer values.  Also, rounding
to int potentially makes the value less normally distributed, which
might interfere with applying t-tests.

There may still be some benefit in reducing precision to save space,
especially when storing results from many builds together, but that
reduction of precision can be done in a different layer.

Bug: 68471
Bug: 74469
Test: fx test host_x64/summarize_perf_results_test
Change-Id: I10cdc888bd38d2c928f1a4210f49bc6512dcb9d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517682
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Reviewed-by: Nathan Rogers <nathanrogers@google.com>