[docs] Fix "Formerly known as FTP-NNN" notices
This CL edits RFC-0026 and RFC-0036 to add "Formerly known as FTP-NNN"
notices, which were missing. It also moves these notices to the top of
the page, below "## Rejection rationale" instead of below "## Summary",
for all rejected RFCs. It used to be inconsistent, with 7 rejected RFCs
using the former style and 4 RFCs (0031, 0034, 0038, and 0044) using the
latter. Now all 11 use the former.

Change-Id: I8b7fcfef1e09085fe1f0ccf05b79b1fbb5737de8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517141
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>