[volume_image]: Fix build time regression.
The previous change that caused the bug was fixing a problem where the
decompressor could provide a bigger decompressed buffer than what the
caller would expect. This is fixed, but the implementation relied on
this for the buffer to expand.

In the end this caused that the buffer never expanded because it was
never set. So decompressing 19 bytes a time (19 cause its the minimum).

The decompressor has a default buffer size of 64 Kb, tried multiple
buffer sizes, but due to the implementation of the fvm_sparse_image
reader, it doesnt provide much help. (A lot of vector resizing
happening).

Updated callsites, effectively this make it so there is an upperbound on
the expected decompression data, which is useful for streaming
the decompressed data on a compressed source into a destination with
tighter memory consumption.

```
time /usr/bin/env ../../build/gn_run_binary.sh
../../prebuilt/third_party/clang/linux-x64/bin host_x64-asan-ubsan/fvm obj/build/images/fvm.fastboot.blk
ftl-raw-nand --nand-page-size 4096 --nand-oob-size 8 --nand-pages-per-block 64 --nand-block-count 1564
--max-disk-size 410746880 --sparse ./obj/build/images/fvm.blob.sparse.blk --reserve-slices 32

real    0m19.950s
user    0m19.068s
sys     0m0.883s
```

Bug: 74696 #Fixed
Test: storage-volume-image-test, fvm-sparse-image-reader-test
Change-Id: I1930d8cab11f00d9024c2bbff5d1e591b228015b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517745
Reviewed-by: Josh Gargus <jjosh@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>