[kernel] Migrate kernel.entropy-test.*
Added a constant to defined the default max, and wrapped the length on a
function that returns the macro value when defined for the test.

`quality_test.cc` was bit rotten, had to update it so it could build.

Created a bug to either add a build only bot or remove it:
fxb/74705

Bug: 53594
Change-Id: I96e88a1a31d63b5b147d3c21c8d1e4898ca2f0fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516781
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>