[build][run_test_component] Remove test_package() instance.
This CL replaces a single test_package() instance into a set
of fuchsia_test_component(), fuchsia_component() and
fuchsia_test_package() targets.

Note that one test binary is moved from executable() to test()
to ensure that they are installed under test/, instead of bin/,
as expected by the current component manifests.

Also note that one ParseArgs test case in run_test_component_unittests
had to be removed, because it relied on the fact that the components of
the run_test_component_test package are listed in the global component
index, which was no longer true (the global component index is
deprecated, and fuchsia_package() does not call add_to_index() as
package() used to), and also probably a bad idea for a unit test anyway.

+ Fix logging-component's component name to match its
  manifest name (i.e. "logging_component" instead of
  "logging-component"), otherwise the test would break.

Bug: 65068
Change-Id: I8cb44639d73e17497a4a46becc4b7a37ab13b26f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514381
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Shai Barack <shayba@google.com>