[fxfs] Implement unlink
For now, files are always moved to the graveyard. Eventually we can
optimize this for the case where the file has no open references by
immediately deleting the file, but that requires appropriate locking on
the file node which isn't implemented yet.

For now, we don't ever purge files which are placed into the graveyard,
so they'll continue to take up space until we implement that.

Also fixed a small TODO in SimpleAllocator (to not assume an infinite
device), since this was tripped over when writing a test.

Change-Id: I904e2bb26ecc18ab5bb1763158cd56be9d708691
Tested: unit.
Multiply: fxfs-tests: 100
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514621
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>