[blobfs] Use shared reads.
Utilizes shared locks for reads which allows multiple read operations
to execute at the same time. This has no effect on the current code
because there is only one fidl thread, and the current (old) pager
doesn't lock with the blob at all. But in the new pager this will allow
potentially more than one pager and/or fidl thread.

The shared locking is also necessary to allow the fidl-based read
operations to be backed by a page request, since this re-enters the
lock.

The paged/unpaged variants of the BlobfsCheckerTest are removed. The
flag was no longer hooked up and these tests were the same.

The BlobCkeckerTests are updated to use the new BlobfsTestSetup class
which does additional work required for the new pager. These tests now
pass in the new pager.

Change-Id: I3408eb5db01a5c2b42e26729eaa62c948ccd0562
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517007
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>