[netstack] Poll slower in reachability test
Change the reachability integration test to poll inspect data every
100ms rather than as fast as possible.

This will hopefully mitigate against flake where sometimes when
running in CQ, an attempt to get inspect data will return the
"InconsistentSnapshot" error. This is usually encountered when a
component makes frequent writes (on the order of 1k writes/s), causing
cache coherency delays. It is suspected that some combination of
polling as fast as possible and the CQ environment is causing this
infrequent flake even though reachability doesn't write very
frequently, but due to polling as fast as possible increasing the
chances that a read attempt will coincide with the component writing
inspect data.

Bug: 71452
Change-Id: I08950d1a089911ae95b04677997e50e91b5b44a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514920
Commit-Queue: Tony Gong <gongt@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>