[fake-clock] Use async task and cancellation
The future attached to an fasync::Task stops being polled when the
handle is dropped, effectively cancelling its execution. This is
guaranteed to stop execution in fake clock as only a single thread is
present, and has the advantage that the code cancelling the task does
not need to be async.

Testing: existing tests

Change-Id: I1270c9d08e4b753c816078e6ba5eeeee9b4f0867
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517903
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>