[hypervisor] Remove dependency on Zircon utc_offset.
Zircon no longer maintains a UTC clock and, per RFC-0008, the syscalls
and internal state associated with the old clock will soon be removed
entirely which means we need an alternative solution for hypervisor.

I synced with Abdulla offline and he's in support of removing the usage
of utc_offset given guest OSs have other options for setting their UTC
and noting we could find another way of setting UTC from the userspace
clock in the future if necessary.

This CL is very much a barebones implementation of the removal; none of
the structs are changed, the boot time is fixed at zero and the offset
is just set to monotonic time. I'm not familiar with hypervisor so very
willing to take alternative suggestions if you'd like anything done
different.

Bug: 74471
Change-Id: Ied51d79c6ba226122c14cf5d7a81cab70cd54e09
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516330
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>