[fshost] remove registry
Nothing was using the registry for anything, and we have no plans to use
it in the future. It was originally meant as a new way of routing
filesystem apis. This role is now taken by component_manager.

Fixed: 75028
Test: fx test fshost-tests
Change-Id: I09dd7e23bdc89bcd4bcbd14544d1a9edd7b47370
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510507
Commit-Queue: Stephen Demos <sdemos@google.com>
Reviewed-by: Manali Bhutiyani <manalib@google.com>
API-Review: Manali Bhutiyani <manalib@google.com>