[tests][e2e][intl] Runs the timezone{,-flutter}-test dart server in a test environment.
The old code ran the timezone dart FIDL server (used to hoist
information from a Dart or Flutter VM) outside of a test environment.

This left the dart or flutter runner running even after the test
terminated, which seems to impact our test infrastructure somehow, by
blocking the entire fixture.

The updated code sets up a test environment based on the example at:
https://cs.opensource.google/fuchsia/fuchsia/+/master:src/connectivity/network/netstack3/tools/enclosed_runner/src/main.rs;l=141?q=add_proxy_service&ss=fuchsia%2Ffuchsia:src%2F
(thanks to anmittal@ for a pointer) to run this FIDL server in.

As result, at the test termination, when the test environment is torn
down, both the timeserver and the runner are terminated.

As an additional result, though, the component manifest for these tests
now have to connect to a bunch of additional services which used to be
available through the environment but now need to be explicitly routed
in and out of the test environment.  This required a change both to the
component manifest *and* the test fixture code itself, since the test
environment FIDL protocol wiring is done in code.

This makes the test a bit unwieldy to follow and reason about, but
that's apparently what we must do to ensure this works properly.

Multiply: timezone-test,timezone-flutter-test
Fixed: 73859
Change-Id: I41f760516303c8cdb830b5555200a2d9be804441
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518740
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>