[build][x86][gcc] add a few feature bits to bring it up to parity with clang
With the recent -march=x86-64-v2 change for clang, the gcc build got
left behind slightly. Move a now redundant switch (-mcx16) under gcc and
add the few features that the kernel might use (-mpopcnt, -msahf) to
gcc.

Confirmed that gcc builds compile and run, and the popcnt instruction is
used.

This CL isn't really long for the world, since it appears that upcoming
gcc 11 will have a matching x86-64-v2 march, but the work and testing is
already done so may as well put this in anyway.

Change-Id: I8254de5107409c07bb37017f8c67361954319505
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518745
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>