[kernel][lib/arch] Make arch::InitializeBootCpuid idempotent
We have a use-case for repopulating our CPUID cache and,
absent-mindedly, a second call to arch::InitializeBootCpuid was
introduced to accomplish this. However, in the function's current form,
it just overwrites the cache with nonsense. In this change, we update
the layout of arch::CpuidIo to store the associated leaf and subleaf
and update the population logic to allow for the function to be called
multiple times.

Bug: 74527
Change-Id: I69fe36d84adec2b703fa74f9e415901ecbbaafa5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515801
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Steve Austin <steveaustin@google.com>