[kernel][lib/arch] Introduce arch::ApplyX86ErrataWorkarounds()
This change introduces a consolidated
arch::ApplyX86ErrataWorkarounds(), which also refactors the existing
logic specific to AMD family 0x17, models 0x00-0x0f. The intent is that
this function will be extended for other AMD and Intel processors.

We removed the workaround for erratum 1090, as that was sourced from an
NDA-guarded Revision Guide - meaning we were unable to document what it
was doing and why - and as it did not seem critical to mitigate neither
(Linux nor FreeBSD reference it).

Bug: 61093
Change-Id: Ibde4a1240ebfcc18daca3b505778b5449c82b474
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515360
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>