[artifactory] Upload objs_to_refresh_ttl.txt and set CustomTime on uploads.
This will list the blobs and debug binaries that we skipped uploading
because they already existed in the GCS bucket. A post-processor can
then use this list to determine which objects to update timestamps for
so that they don't expire.
This also sets the CustomTime attribute on all uploads so that
we can change our buckets to expire objects based on their CustomTimes
instead of their creation date.

Bug: 66462
Change-Id: I9f5e1745d218a1c278bb2fb15dce405957ec3cda
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516680
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>