[input] Fixes some issues related to key ordering
The whole story is a bit involved, please see fxbug.dev/74271 for
a run-down of issues.

The short of it is that the current delivery discipline around
simultaneous events may cause a modifier key to be applied to a key
that preceded it.  It is a side effect of a few design decisions at
different points in the input pipeline and the test fixtures, each
of which quite fine on its own, but when strung together result in
an unwanted behavior.

This change cleans up a few issues mentioned in the bug, and may
incidentally fix some.

This is a prerequisite change to fxr/513823.

Tested: fx ... lsc ...
Multiply: experiences_ermine_session_shell_e2e_test
Bug: 74271
Change-Id: I8f31202c1184b6051ba204a828138bc887635407
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514048
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Alice Neels <neelsa@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>