[netemul] Implement `ChildUses.all`
When a component in a test realm requests to `use` "all" services, as
opposed to enumerating particular ones, it gets every built-in netemul
service, as well as all services provided by other components in the
same test realm.

In order to test this, this CL introduces a `ConnectToService` method on
the `fuchsia.netemul.test/Counter` protocol, and exposes the `Counter`
service from the `counter` component twice, aliased as `CounterA` and
`CounterB` -- this allows two identical `counter` components to exist in
the same test realm and both expose the same underlying service, but
without duplicate service names.

Fixed: 73452, 74817
Test: fx test -d netemul-v2-tests
Change-Id: I34a76c2801a5cdd4ec14817e07ec868567166a11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516461
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>