[overnet] Flake fix
The connection from the process running `onet list-links` to `ascendd`
may not be formed nor be in the process of forming when `onet
list-links` polls itself to get a list of links.

This caused a flake in a test that was measuring how many links were
reported.

Since we know that there must be at least two nodes on a mesh for `onet
list-links` to work (the other being ascendd) we also know that there
must be at least one link per process and so it becomes an easy thing to
just wait for things to settle out before proceeding.

BUG: 73091
Change-Id: I69c82fa8598ae5763f74f008e7d9811cc72e1c9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518746
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Craig Tiller <ctiller@google.com>