[media][tests] include source_set in all-tree compile
Source_set "hermetic_pipeline_test_main" is currently only consumed
by build targets in the vendor/google repo, so it would not be
included in a vanilla top-level build of the entire fuchsia tree.
This CL adds it to the container "//src/media/audio/lib/test:test"
which exists for exactly this purpose.

Bug: 58162

Change-Id: Iaa22138c5a62f3348cf05226e1d6f5afff975647
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519160
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>