[ffx][repository] Bring repository structure plugin in line with rest of tree
There's a convention for subcommand structure in ffx that gives us the
help command automatically as well as a few other niceties. This ports
the framework of the repository plugin to it.

We also add a stub for the list command in addition to the one for the
serve command.

Change-Id: Ic85f42b0c032ffb1e064f49ebfa0a81eac82ecda
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518911
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>